refactor(server): organize llm package
